changes to run with NMR downloaded dataset

+++ app/main.py
@@ -71,24 +71,46 @@ async def upload_images(data: MultiImageData,background_tasks: BackgroundTasks):
 
 @app.post("/reconstruct")
 async def upload_images(data: Parameters,background_tasks: BackgroundTasks):
-    # Update the status dictionary
-    status = {
-        "id": data.id,
-        "progress": f"0 of {data.epochs} epochs",
-        "status": "training"
-    }
-    result = {
-        "id": data.id,
-        "status": "Pending Result"
-    }
     # Specify the filename
     status_f = f'{data.id}.json'
-    result_f = f'{data.id}r.json'       
-    # Write the updated data to the JSON file
-    with open(status_f, 'w') as file:
-        json.dump(status, file)
-    with open(result_f, 'w') as file:
-        json.dump(result, file)
+    result_f = f'{data.id}r.json'
+
+    # Attempt to parse the JSON string
+    try:
+        with open(status_f, 'r') as file:
+            old_data = json.load(file)
+    except json.JSONDecodeError:
+        print(f"Warning: Invalid JSON in {status_f}. Skipping parsing.")
+        old_data = {}
+    
+    # Check if the progress is Failed
+    if old_data.get("progress") == "Failed":
+        # Define the data
+        status = {
+            "id": data.id,
+            "progress": "Failed",
+            "status": "Reconstruction unvailable without colmap"
+        }
+        # Write the data to the JSON file
+        with open(status_f, 'w') as file:
+            json.dump(status, file)
+        return {"message": "Reconstruction unvailable without colmap"}
+    else:
+        # Define the data
+        status = {
+            "id": data.id,
+            "progress %": f"0 of {data.epochs} epochs",
+            "status": "training"
+        }
+        result = {
+            "id": data.id,
+            "status": "Pending Result"
+        }
+        # Write the data to the JSON file
+        with open(status_f, 'w') as file:
+            json.dump(status, file)
+        with open(result_f, 'w') as file:
+            json.dump(result, file)
 
     #execute train.py using subprocess
     train_script_path = 'train.py'
@@ -128,21 +150,21 @@ def process_image(data_images,data_id):
         #filename = f'image_{int(ts)}.jpg'
         filename = f'image_{image_data.url.split("/")[-1]}_{int(ts)}.jpg'
         file_path = os.path.join(f'dataset/{data_id}/dense/images', filename)
-        # Convert base64 string to bytes
-        #urllib.request.urlretrieve(image_data, file_path)
-            # Download the image from the URL
-        response = urllib.request.urlopen(image_data.url)
+        # Download the image from the URL
+        # Create a request with a user-agent header
+        headers = {'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3'}
+        req = urllib.request.Request(image_data.url, headers=headers)
+
+        # Send the request and retrieve the response
+        response = urllib.request.urlopen(req)
         image_bytes = response.read()
         
         # Save the image bytes to the file path
         with open(file_path, 'wb') as file:
             file.write(image_bytes)
-        
-        # Wait for 2 seconds
+
         time.sleep(1)
-            
-        # Optionally, perform background tasks here
-        # background_tasks.add_task(some_background_task, image_bytes)
+    
 
     print("Images saved successfully!")
 
@@ -158,16 +180,35 @@ def process_image(data_images,data_id):
 
     # Run COLMAP sparse reconstruction on the specified folder
     colmap_sparse_reconstruction(image_folder_path, output_sparse_path)
-    rename_folder(f'dataset/{data_id}/dense/0', f'dataset/{data_id}/dense/sparse')
-    print("Sparse reconstruction completed successfully!")
-    # Update the status dictionary
-    status = {
-        "id": data_id,
-        "progress": "Done",
-        "status": "Ready for reconstruction"
-    }
-    # Specify the filename
-    status_f = f'{data_id}.json'       
-    # Write the updated data to the JSON file
-    with open(status_f, 'w') as file:
-        json.dump(status, file)
+    old_sparse_path = f'dataset/{data_id}/dense/0'
+    new_sparse_path = f'dataset/{data_id}/dense/sparse'
+    rename_folder(old_sparse_path, new_sparse_path)
+
+    if os.path.exists(os.path.join(new_sparse_path, 'cameras.bin')):
+        print("Sparse reconstruction completed successfully!")
+        # Update the status dictionary
+        status = {
+            "id": data_id,
+            "progress": "Done",
+            "status": "Ready for reconstruction"
+        }
+        # Specify the filename
+        status_f = f'{data_id}.json'       
+        # Write the updated data to the JSON file
+        with open(status_f, 'w') as file:
+            json.dump(status, file)
+    else:
+        print("Sparse reconstruction completed successfully!")
+        # Update the status dictionary
+        status = {
+            "id": data_id,
+            "progress": "Failed",
+            "status": "Colmap failed to reconstruct"
+        }
+        # Specify the filename
+        status_f = f'{data_id}.json'       
+        # Write the updated data to the JSON file
+        with open(status_f, 'w') as file:
+            json.dump(status, file)
+        print("Sparse reconstruction failed.")
+
